Fixed build error introduced with r448122 about DataObjectList.append not supporting int type
git-svn-id: https://svn.apache.org/repos/asf/incubator/tuscany/cpp/sca@448124 13f79535-47bb-0310-9956-ffa450edef68

+++ runtime/extensions/ws/service/axis2c/src/tuscany/sca/ws/WSServiceProxy.cpp
@@ -381,7 +381,7 @@ namespace tuscany
                             }
                         case Operation::INT: 
                             {
-                                l.append(*(int*)operation.getReturnValue());
+                                l.append(*(long*)operation.getReturnValue());
                                 break;
                             }
                         case Operation::LONG: 
@@ -391,12 +391,12 @@ namespace tuscany
                             }
                         case Operation::USHORT: 
                             {
-                                l.append(*(short*)operation.getReturnValue());
+                                l.append(*(unsigned short*)operation.getReturnValue());
                                 break;
                             }
                         case Operation::UINT: 
                             {
-                                l.append(*(int*)operation.getReturnValue());
+                                l.append(*(long*)operation.getReturnValue());
                                 break;
                             }
                         case Operation::ULONG:
